Raise error when default configs have too wide permissions
