+++ tests/test_main.py
@@ -141,3 +141,21 @@ def test_if_there_are_no_option_duplicates(runner):
     _check(ctx.command)
 
     assert duplicates == {}, "\n".join(duplicates)
+
+
+def test_fail_command_when_default_config_has_too_wide_permissions(
+    snowflake_home: Path,
+):
+    from snowflake.cli.app.cli_app import app
+
+    runner = CliRunner()
+
+    config_path = snowflake_home / "config.toml"
+    config_path.touch()
+    config_path.chmod(0o777)
+
+    result = runner.invoke(app, ["sql", "-q", "select 1"])
+
+    assert result.exit_code == 1, result.output
+    assert result.output.__contains__("Error")
+    assert result.output.__contains__("Configuration file")
